python_service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-exporter-otlp==1.21.0
grpcio==1.59.3
grpcio-tools==1.59.3 